Enhance and test security

+++ trashtalk/views/cleanups.py
@@ -8,7 +8,7 @@ from geopy.exc import GeopyError
 from trashtalk.google_sheets import send_to_sheet
 from trashtalk.seeclickfix import postSCFix
 from trashtalk.factories import cleanup_factory, location_factory
-from trashtalk.models import Cleanup, db_session  # User
+from trashtalk.models import Cleanup, db_session, User
 from trashtalk.html_constants import HtmlConstants
 from trashtalk.input_handling import *
 from trashtalk.utils import get_location
@@ -49,7 +49,6 @@ def get(cleanup_id):
     :param cleanup_id:
     :return: show template
     """
-
     cleanups = db_session.query(Cleanup).get(cleanup_id)
     # Method sort those logged in and out, in reference to participating in cleanups
     if current_user.is_authenticated:
@@ -69,6 +68,7 @@ def get(cleanup_id):
 
 
 @bp.route('/new')
+@login_required
 def new():
     """
     Render template form for creating a new `Cleanup`. POSTs to cleanups/create
@@ -94,24 +94,30 @@ def edit(cleanup_id):
     :return: edit template.
     """
     # TODO: Update form to include current values for current cleanup
+
     cleanups = db_session.query(Cleanup).get(cleanup_id)
-    return render_template('cleanup/edit.html',
-                           section='Edit Cleanup',
-                           id=cleanups.id,
-                           date=cleanups.date,
-                           current_address=cleanups.location.number,
-                           start_time= hour_min_value(cleanups.start_time),
-                           start_time_of_day = am_pm_value(cleanups.start_time),
-                           end_time = hour_min_value(cleanups.end_time),
-                           end_time_of_day = am_pm_value(cleanups.end_time),
-                           description=cleanups.description,
-                           date_pattern = html_constants.date_pattern,
-                           date_placeholder = html_constants.date_placeholder,
-                           time_pattern = html_constants.time_pattern,
-                           time_placeholder = html_constants.time_placeholder)
+    if cleanups.host_id == current_user.id: #Only let host edit cleanup
+        return render_template('cleanup/edit.html',
+                               section='Edit Cleanup',
+                               id=cleanups.id,
+                               date=cleanups.date,
+                               current_address=cleanups.location.number,
+                               start_time= hour_min_value(cleanups.start_time),
+                               start_time_of_day = am_pm_value(cleanups.start_time),
+                               end_time = hour_min_value(cleanups.end_time),
+                               end_time_of_day = am_pm_value(cleanups.end_time),
+                               description=cleanups.description,
+                               date_pattern = html_constants.date_pattern,
+                               date_placeholder = html_constants.date_placeholder,
+                               time_pattern = html_constants.time_pattern,
+                               time_placeholder = html_constants.time_placeholder)
+    else:
+        flash("You do not have rights to edit this cleanup")
+        return redirect(url_for('cleanups.get', cleanup_id=cleanup_id))
 
 
 @bp.route('/<int:cleanup_id>', methods=['PUT', 'DELETE'])
+@login_required
 def update(cleanup_id):
     """
     Create, update or delete a `Cleanup`
@@ -121,32 +127,39 @@ def update(cleanup_id):
     """
     cleanup_dict = request.form.to_dict()
     method = cleanup_dict['method']
+
     cleanups = db_session.query(Cleanup).get(cleanup_id)
-    if method == 'POST': #PUT
-        current_app.logger.debug("REQUEST FORM: %s", request.form)
+    if cleanups.host_id == current_user.id:  # Only let host edit cleanup
 
-        start_time = twenty_four_time(cleanup_dict['start_time'],cleanup_dict['start_time_of_day'])
-        end_time = twenty_four_time(cleanup_dict['end_time'],cleanup_dict['end_time_of_day'])
-        cleanup_dict.update({"start_time": start_time, "end_time": end_time})
+        if method == 'POST': #PUT
+            current_app.logger.debug("REQUEST FORM: %s", request.form)
 
-        full_address=get_full_address(cleanup_dict)
-        try:
-            geoloc = get_location(full_address)
-        except:
-            current_app.logger.exception("There was an error finding location.")
-            flash("s%: Address not found" % full_address)
-            return redirect(url_for('cleanups.edit', cleanup_id=cleanup_id))
-        else:
-            location_dict = {"number": geoloc.address, "latitude": geoloc.latitude, "longitude": geoloc.longitude}
-            cleanup_dict.pop("location",None)#Hack- location should be renamed, coming from HTML
-            cleanups.update(**cleanup_dict)
-            cleanups.location.update(**location_dict)
-            return redirect(url_for('cleanups.get', cleanup_id=cleanup_id))
+            start_time = twenty_four_time(cleanup_dict['start_time'],cleanup_dict['start_time_of_day'])
+            end_time = twenty_four_time(cleanup_dict['end_time'],cleanup_dict['end_time_of_day'])
+            cleanup_dict.update({"start_time": start_time, "end_time": end_time})
 
-    elif method == 'DELETE':
-        print("now deleting")
-        cleanups.delete()
-        return redirect(url_for('cleanups.get_all'))
+            full_address=get_full_address(cleanup_dict)
+            try:
+                geoloc = get_location(full_address)
+            except:
+                current_app.logger.exception("There was an error finding location.")
+                flash("s%: Address not found" % full_address)
+                return redirect(url_for('cleanups.edit', cleanup_id=cleanup_id))
+            else:
+                location_dict = {"number": geoloc.address, "latitude": geoloc.latitude, "longitude": geoloc.longitude}
+                cleanup_dict.pop("location",None)#Hack- location should be renamed, coming from HTML
+                cleanups.update(**cleanup_dict)
+                cleanups.location.update(**location_dict)
+                return redirect(url_for('cleanups.get', cleanup_id=cleanup_id))
+
+        elif method == 'DELETE':
+            print("now deleting")
+            cleanups.delete()
+            return redirect(url_for('cleanups.get_all'))
+
+    else:
+        flash("You do not have rights to edit this cleanup")
+        return redirect(url_for('cleanups.get', cleanup_id=cleanup_id))
 
 
 @bp.route('/create', methods=["POST"])
@@ -189,7 +202,7 @@ def create():
         return redirect(url_for('cleanups.get', cleanup_id=new_cleanup.id))
 
 
-@bp.route('/<int:cleanup_id>/delete', methods=["POST"])
+@bp.route('/join', methods=["POST"])
 @login_required
 def join():
     """
@@ -203,7 +216,7 @@ def join():
     return redirect(url_for('cleanups.get', cleanup_id=cleanup_id))
 
 
-@bp.route('/join', methods=["POST"])
+@bp.route('/leave', methods=["POST"])
 @login_required
 def leave():
     """
@@ -227,14 +240,18 @@ def send_to_scf(id):
     :return:
     """
     cleanup = db_session.query(Cleanup).filter(Cleanup.id == id).first()
-    current_app.logger.debug("Address: %s" % cleanup.location.number)
-    api_request = postSCFix(cleanup)
-    response = api_request.json()  # Contains Response from SeeClickFix
-    cleanup.html_url = response['html_url']  # Add to SQL Database
-    db_session.add(cleanup)
-    db_session.commit()
-    return redirect(url_for('cleanups.get', cleanup_id=id))
-
+    cleanups = db_session.query(Cleanup).get(id)
+    if cleanups.host_id == current_user.id:  # Only let host edit cleanup
+        current_app.logger.debug("Address: %s" % cleanup.location.number)
+        api_request = postSCFix(cleanup)
+        response = api_request.json()  # Contains Response from SeeClickFix
+        cleanup.html_url = response['html_url']  # Add to SQL Database
+        db_session.add(cleanup)
+        db_session.commit()
+        return redirect(url_for('cleanups.get', cleanup_id=id))
+    else:
+        flash("You do not have rights to edit this cleanup")
+        return redirect(url_for('cleanups.get', cleanup_id=id))
 
 @bp.route('/public_works/<id>', methods=['GET'])
 @login_required
@@ -264,11 +281,16 @@ def send_public_works(id):
     :param id: cleanup id
     :return:
     """
-    tool_data=request.form.copy()
-    send_to_sheet(id, tool_data)  # Very slow function
-    cleanup = db_session.query(Cleanup).get(id)
-    cleanup.notified_pw=True
-    db_session.add(cleanup)
-    db_session.commit()
+    cleanups = db_session.query(Cleanup).get(id)
+    if cleanups.host_id == current_user.id:  # Only let host edit cleanup
+        tool_data=request.form.copy()
+        send_to_sheet(id, tool_data)  # Very slow function)
+        cleanups.notified_pw=True
+        db_session.add(cleanups)
+        db_session.commit()
 
-    return redirect(url_for('cleanups.get', cleanup_id=id))
+        return redirect(url_for('cleanups.get', cleanup_id=id))
+
+    else:
+        flash("You do not have rights to edit this cleanup")
+        return redirect(url_for('cleanups.get', cleanup_id=id))
